moved Path.java for consistency
